+++ module/resources/views/admin/details.blade.php
@@ -1,5 +1,5 @@
 <x-layouts.app-shell :title="__('module_cleaner::messages.details.title')" :subtitle="__('module_cleaner::messages.details.subtitle')">
-    <div class="settings-layout">
+    <div class="settings-category-layout">
         @include('module_cleaner::admin.partials.section-nav')
 
         @php