+++ module/resources/views/admin/registry.blade.php
@@ -1,5 +1,5 @@
 <x-layouts.app-shell :title="__('module_cleaner::messages.registry.title')" :subtitle="__('module_cleaner::messages.registry.subtitle')">
-    <div class="settings-layout">
+    <div class="settings-category-layout">
         @include('module_cleaner::admin.partials.section-nav')
 
         @if ($errors->any())